Remove the thread switching between the SocketReaderNode and the RtpDecoderNode
A design flaw in thread switching can cause latency in thread scheduling when other applications on the AP side are congested. To eliminate this latency, the ImsMedia RX stream has been modified to avoid switching threads between the SocketReaderNode, which receives data frames from the socket, and the RtpDecoderNode, which performs RTP depacketization.

1. Remove the thread switching between the SocketReaderNode and the RtpDecoderNode
2. Stop the node before update in the AudioStreamGraphRtpRx

Bug: 304901905
Bug: 308062939
Test: Verified by the VoWifi call in ATT 19093 certification test. atest
ImsMediaNativeTests

Change-Id: I601629d027a3d27be8a9fdd165782f0407d8ef5b