Reduce the JB delay and fix the JB size calculation logic
The audio delay needs to be optimized by failing the total delay in certification test. There are some point to reduce the delay in the jitter buffer side. I change the logic to reduce the delay in two points.

Delay reduction
1. Reduce max 10ms in wait time in JB data pulling
2. Change roundup from 20ms to 10ms in calculate the JB size to reduce 10ms in playout delay

Fix the bug
1. Fix the too short or long JB size by normalization of the accumulated jitter
2. Remove not decrease the jitter buffer delay condition in SID period

Improvement
1. Remove the condition set 200ms in JB size increasing interval threshold to update the jitter buffer size every packet received
2. Change the jitter buffer logs for easy debugging

Bug: 309548135
Bug: 312619696
Test: VoWifi call test in live network, verified in ATT audio handset
certification test cases, atest ImsMediaNativeTests

Change-Id: Ib3e33ea9238c763f268be4eddc44adf16fead160