Fix JBM to start as initial state when the ssrc changing called
Problem: There are more packet drops than usual after a handover from VoLTE to VoWiFi.
When the HO resets the JBM, it changes its state to "wait" for a specific time (60ms) based on the initial buffer size. However, the actual buffer size remains larger (140ms) due to not resetting it.

Fix to this, reset the current JB size when the handover occured to match the estimation and the buffering time.

Bug: 315557170
Test: Verified in live network, atest ImsMediaNativeTests
Change-Id: I72f5bc4f00bc0ab2b1236e01014bfe3108b7fa60