Fix the overflow in ImsMediaTimer
When the GetTimeInMilliSeconds method within the ImsMediaTimer class is invoked, a potential overflow issue arises. The method attempts to convert a 64-bit counter, representing microseconds, into a 32-bit value in milliseconds by dividing it by 1000. However, this conversion can lead to overflow if the counter reaches the maximum value that a 32-bit unsigned integer can hold.

To circumvent this overflow and ensure accurate time differences in milliseconds, a modification has been implemented. The corrected approach involves obtaining the remainder of the counter value when divided by the maximum unsigned integer. This effectively prevents overflow and guarantees accurate millisecond calculations.

Bug: 316822152
Test: atest ImsMediaNativeTests:ImsMediaTimer with iteration 400 times
Change-Id: Ia665f028f518315b060dd0c08ff705e29e622ae4