Use timed mutex to timeout and break from deadlock.
If a deadlock occurs during a call, it causes successive call failures as ImsMedia service can't process any new requests. Currently, the only way to recover is device reboot. This improvement will introduce a timeout for the mutex lock. If the lock isn't released within a certain time, the ImsMedia service will automatically restart. The current call will be dropped, which is the same behavior as before. However, this prevents the issue from spreading to subsequent calls, so they will still work.

Bug: 342326367
Bug: 489373581
Flag: NONE <AOSP migration>
Test: Stress testing in live network, atest ImsMediaNativeTests
Change-Id: I460fbe8d375a7a492ea505305a4de7ed3e9ce07c